fingerprint_tls::tls_extensions

Trait TLSExtension

pub trait TLSExtension:
    Debug
    + Any
    + Send {
    // Required methods
    fn len(&self) -> usize;
    fn read(&self, buf: &mut [u8]) -> Result<usize>;
    fn extension_id(&self) -> ExtensionID;
    fn as_any(&self) -> &dyn Any;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

TLS extension trait Corresponds to Go version’s tls.TLSExtension interface

Required Methods

fn len(&self) -> usize

Get extension length (include header) Corresponds to Go version’s Len() int

fn read(&self, buf: &mut [u8]) -> Result<usize>

Read extension count data to bytes buffer Corresponds to Go version’s Read(p []byte) (n int, err error)

fn extension_id(&self) -> ExtensionID

Get extension ID

fn as_any(&self) -> &dyn Any

Convert to Any trait object, for toward down transform

Provided Methods

fn is_empty(&self) -> bool

Check extension whether as empty Default implement: length as 0 when as empty